+++ openpecha/serializers/__init__.py
@@ -1,6 +1,5 @@
 from .editor import EditorSerializer
 from .epub import EpubSerializer
-from .foot_note import SerializeFootNote
 from .hfml import HFMLSerializer
 from .serializemd import SerializeMd
 from .pedurma import PedurmaSerializer